refactor(chat): split deploy/share state into its own Zustand store
The viewer store was bundling five fields that exclusively belong to
the app share/deploy lifecycle (isSharing, isDeploying,
isTokenDialogOpen, pendingDeployAppId, complexDeployApp) alongside
panel navigation and viewer-lifecycle state they have nothing to do
with. Splits them into apps/web/src/domains/chat/deploy-store.ts —
one store per independent data concern, per CONVENTIONS § State
management.

- New store: useDeployStore with the five fields + seven actions
  (startSharing/finishSharing/startDeploying/finishDeploying/
  showTokenDialog/hideTokenDialog/setComplexDeployApp) and a reset.
  Same direct-named-actions pattern + createSelectors wrapper as
  the existing chat-domain stores.
- chat-route-content subscribes to useDeployStore.use.isSharing()
  and .isDeploying() directly via the .use.* selectors — same
  pattern it already uses for interaction-store and turn-store.
  The viewerState prop is dropped from ChatRouteContentProps now
  that it carries nothing the component still reads.
- chat-page no longer subscribes to deploy state (chat-route-content
  owns its own subscription) and switches its startSharing /
  startDeploying calls to useDeployStore.getState().
- use-app-viewer-actions switches all share/deploy store-action
  calls from useViewerStore.getState() to useDeployStore.getState().
  The hook isn't wired into any caller yet (migration leftover) but
  staying consistent with the new store binding keeps it ready.
- Share/deploy tests move into deploy-store.test.ts; viewer-store
  reset test trimmed accordingly.
